feat: Refactor project architecture: moved models to domain/entities and updated imports
