+++ apps/custom_auth/infrastructure/api/v1/serializers/profile_information_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from apps.custom_auth.models import ProfileInformation
+
+
+class ProfileInformationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ProfileInformation
+        fields = ['about_me', 'disciplines', 'contact_info']
