+++ apps/custom_auth/infrastructure/api/v1/serializers/user_serializer.py
@@ -1,0 +1,43 @@
+from rest_framework import serializers
+from django.contrib.auth.hashers import make_password
+from apps.custom_auth.models import User
+from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
+
+
+class UserListSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ['id', 'first_name', 'last_name', 'username']
+
+
+class UserTokenObtainPairSerializer(TokenObtainPairSerializer):
+    @classmethod
+    def get_token(cls, user):
+        token = super().get_token(user)
+        token['user_id'] = user.id
+        return token
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ['first_name', 'last_name', 'scopus_id', 'institution',
+                  'website', 'investigation_camp', 'profile_picture', 'email_institution']
+
+    def validate_website(self, value):
+        if value and not value.startswith(('http://', 'https://')):
+            value = 'http://' + value
+        return value
+
+
+class RegisterSerializer(serializers.ModelSerializer):
+    password = serializers.CharField(write_only=True)
+
+    class Meta:
+        model = User
+        fields = ['first_name', 'last_name',
+                  'username', 'scopus_id', 'password']
+
+    def create(self, validated_data):
+        validated_data['password'] = make_password(validated_data['password'])
+        return super().create(validated_data)
