+++ apps/custom_auth/infrastructure/api/v1/views/group_views.py
@@ -1,0 +1,22 @@
+from rest_framework import generics, permissions, status
+from rest_framework.response import Response
+from apps.custom_auth.models import Group
+from apps.custom_auth.infrastructure.api.v1.serializers.group_serializer import GroupSerializer
+
+
+class GroupListCreateView(generics.ListCreateAPIView):
+    queryset = Group.objects.all()
+    serializer_class = GroupSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def perform_create(self, serializer):
+        serializer.save(admin=self.request.user)
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        if serializer.is_valid():
+            self.perform_create(serializer)
+            headers = self.get_success_headers(serializer.data)
+            return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
