+++ apps/custom_auth/infrastructure/api/v1/views/post_views.py
@@ -1,0 +1,35 @@
+from rest_framework import generics, permissions
+from rest_framework.parsers import MultiPartParser, FormParser
+from apps.custom_auth.models import Post
+from apps.custom_auth.infrastructure.api.v1.serializers.post_serializer import PostSerializer
+
+
+class PostCreateView(generics.ListCreateAPIView):
+    queryset = Post.objects.all()
+    serializer_class = PostSerializer
+    permission_classes = [permissions.IsAuthenticated]
+    parser_classes = [MultiPartParser, FormParser]
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
+
+
+class PostListView(generics.ListAPIView):
+    serializer_class = PostSerializer
+    permission_classes = [permissions.AllowAny]
+
+    def get_queryset(self):
+        user_id = self.request.query_params.get('user_id')
+        if user_id:
+            return Post.objects.filter(user_id=user_id).order_by('-created_at')
+        return Post.objects.none()
+
+
+class PostDeleteView(generics.DestroyAPIView):
+    queryset = Post.objects.all()
+    serializer_class = PostSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        return queryset.filter(user=self.request.user)
