+++ apps/custom_auth/views.py
@@ -1,167 +1,4 @@
-from rest_framework import generics, status
-from rest_framework.response import Response
-from rest_framework.exceptions import ValidationError
-from rest_framework import generics, permissions
-from rest_framework.exceptions import PermissionDenied
-from rest_framework.parsers import MultiPartParser, FormParser
-from .models import User
-from .serializers import *
-from rest_framework import generics, permissions
-from .models import ProfileInformation
-from .serializers import ProfileInformationSerializer
-from rest_framework.exceptions import PermissionDenied
-
-
-class PostCreateView(generics.ListCreateAPIView):
-    queryset = Post.objects.all()
-    serializer_class = PostSerializer
-    permission_classes = [permissions.IsAuthenticated]
-    parser_classes = [MultiPartParser, FormParser]
-
-    def perform_create(self, serializer):
-        serializer.save(user=self.request.user)
-
-
-class PostListView(generics.ListAPIView):
-    serializer_class = PostSerializer
-    permission_classes = [permissions.AllowAny]
-
-    def get_queryset(self):
-        user_id = self.request.query_params.get('user_id')
-        if user_id:
-            return Post.objects.filter(user_id=user_id).order_by('-created_at')
-        return Post.objects.none()
-
-
-class PostDeleteView(generics.DestroyAPIView):
-    queryset = Post.objects.all()
-    serializer_class = PostSerializer
-    permission_classes = [permissions.IsAuthenticated]
-
-    def get_queryset(self):
-        queryset = super().get_queryset()
-        return queryset.filter(user=self.request.user)
-
-
-class ProfileInformationDetailView(generics.RetrieveUpdateDestroyAPIView):
-    queryset = ProfileInformation.objects.all()
-    serializer_class = ProfileInformationSerializer
-    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
-
-    def get_object(self):
-        profile_information, created = ProfileInformation.objects.get_or_create(
-            user=self.request.user)
-        return profile_information
-
-    def perform_update(self, serializer):
-        if self.request.user.id != serializer.instance.user.id:
-            raise PermissionDenied(
-                "You do not have permission to edit this profile.")
-        serializer.save()
-
-    def perform_destroy(self, instance):
-        if self.request.user.id != instance.user.id:
-            raise PermissionDenied(
-                "You do not have permission to delete this profile.")
-        if not instance.about_me and not instance.disciplines and not instance.contact_info:
-            instance.delete()
-        else:
-            raise ValidationError(
-                "Profile information must be empty to be deleted.")
-
-
-class PublicProfileInformationDetailView(generics.RetrieveAPIView):
-    queryset = ProfileInformation.objects.all()
-    serializer_class = ProfileInformationSerializer
-    permission_classes = [permissions.AllowAny]
-    lookup_field = 'user__id'
-
-
-class GroupListCreateView(generics.ListCreateAPIView):
-    queryset = Group.objects.all()
-    serializer_class = GroupSerializer
-    permission_classes = [permissions.IsAuthenticated]
-
-    def perform_create(self, serializer):
-        serializer.save(admin=self.request.user)
-
-    def create(self, request, *args, **kwargs):
-        print("Data received: ", request.data)
-        serializer = self.get_serializer(data=request.data)
-        if serializer.is_valid():
-            self.perform_create(serializer)
-            headers = self.get_success_headers(serializer.data)
-            return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
-        else:
-            print("Validation errors: ", serializer.errors)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class UserListView(generics.ListAPIView):
-    serializer_class = UserListSerializer
-    permission_classes = [permissions.IsAuthenticated]
-
-    def get_queryset(self):
-        return User.objects.exclude(id=self.request.user.id)
-
-
-class UserUpdateView(generics.UpdateAPIView):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-    permission_classes = [permissions.IsAuthenticated]
-    parser_classes = (MultiPartParser, FormParser)
-
-    def get_object(self):
-        obj = super().get_object()
-        if obj.id != self.request.user.id:
-            raise PermissionDenied(
-                "You do not have permission to edit this user.")
-        return obj
-
-    def update(self, request, *args, **kwargs):
-        kwargs['partial'] = True
-        return super().update(request, *args, **kwargs)
-
-
-class UserTokenObtainPairView(TokenObtainPairView):
-    serializer_class = UserTokenObtainPairSerializer
-
-
-class UserDetailView(generics.RetrieveAPIView):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-
-
-class RegisterView(generics.CreateAPIView):
-    queryset = User.objects.all()
-    serializer_class = RegisterSerializer
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        try:
-            serializer.is_valid(raise_exception=True)
-            self.perform_create(serializer)
-            headers = self.get_success_headers(serializer.data)
-            return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
-        except ValidationError as e:
-            return Response(self.format_errors(serializer.errors), status=status.HTTP_400_BAD_REQUEST)
-        except Exception as e:
-            return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-    def format_errors(self, errors):
-        custom_errors = {}
-        for field, field_errors in errors.items():
-            custom_errors[field] = [self.get_custom_error_message(
-                field, error) for error in field_errors]
-        return {'errors': custom_errors}
-
-    def get_custom_error_message(self, field, error):
-        custom_messages = {
-            'username': {
-                'user with this username already exists.': "User with this EMAIL already exists."
-            },
-            'scopus_id': {
-                'user with this scopus id already exists.': "User with this SCOPUS ID already exists."
-            }
-        }
-        return custom_messages.get(field, {}).get(error, error)
+from apps.custom_auth.infrastructure.api.v1.views.group_views import GroupListCreateView
+from apps.custom_auth.infrastructure.api.v1.views.post_views import PostListView, PostCreateView, PostDeleteView
+from apps.custom_auth.infrastructure.api.v1.views.profile_information_views import ProfileInformationDetailView, PublicProfileInformationDetailView
+from apps.custom_auth.infrastructure.api.v1.views.user_views import UserListView, UserUpdateView, UserTokenObtainPairView, UserDetailView, RegisterView
